requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
branca==0.8.2
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
folium==0.20.0
fonttools==4.60.1
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.8.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pydeck==0.9.1
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
streamlit-folium==0.25.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xyzservices==2025.4.0
pytest==8.4.2

# Database dependencies (for database/load_gtfs_to_db.py)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
